+++ NEWS.md
@@ -3,6 +3,7 @@
   * `packages` will now install from GitHub when the package name contains a '/'.
   * `packages` will now take `verbose = TRUE` to announce the installations.
   * Add test coverage for `packages` and `load_package`.
+  * Deprecate space separation for `packages` (instead, must pass multiple arguments).
 
 # Version 0.2.1